M3/W4/crawler/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
db-dtypes==1.0.1
decorator==5.1.1
frozenlist==1.3.0
fsspec==2022.3.0
gcsfs==2022.3.0
google-api-core==2.7.3
google-auth==2.6.6
google-auth-oauthlib==0.5.1
google-cloud-bigquery==3.1.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
greenlet==1.1.2
grpcio==1.46.3
grpcio-status==1.46.3
idna==3.3
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.5
proto-plus==1.20.5
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2